% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Many_hosts_many_parasites_dynamics.R
\name{manyhost_manyparasite_dynamics}
\alias{manyhost_manyparasite_dynamics}
\title{Many hosts, many parasites simulation.}
\usage{
manyhost_manyparasite_dynamics(field.size = 100^2,
  interaction.matrix = matrix(c(1, 1, 2, 2), 2, 2, dimnames =
  list(c("Host 1", "Host 2"), c("Parasite 1", "Parasite 2"))),
  host.number = c(2000, 2000), parasite.number = c(1, 1), gens = 100,
  reps = 5)
}
\arguments{
\item{field.size}{size of host parasite matrix to play on. Must be square.}

\item{interaction.matrix}{the selective advantage to a parasite of landing on a host. Supply a matrix of selection coefficients for each parasite on each host. Hosts on rows, parasites on columns! See examples.}

\item{host.number}{number of hosts to seed the matrix with. Supply a vector of values.}

\item{parasite.number}{number of parasites to seed the matrix with. Supply a vector of values.}

\item{gens}{number of generations to run the model for.}

\item{reps}{how many replicates to run this}
}
\value{
a list of the parasite dynamics, spatial structure and the host matrix?
}
\description{
Take a host grid containing a number of hosts with some benefit to the parasite (randomly spaced, but fixed).
Then overlay a parasite grid contaning some parasites (randomly spaced).
If hosts and parasites co-occur, the parasite will reproduce. Cumulative offspring then randomly overlaid on grid.
}
\details{
Model assumes (1) that hosts are fixed in position and perennial, and are unaffected by parasitism.
  (2) no effect on parasite of nearby hosts, (3) parasites seed randomly over the field, (4) all parasites survive to end of generation,
  (5) parasites do not produce any offspring without a host, and all parasites with a host reproduce and (6) no host independent mortality of parasites
}
\examples{

# say I have two host species and two parasite species.
int.mat <- matrix(c(1,1,2,2), 2, 2, dimnames = list(c("C.cristatus", "L.corniculatus"), c("E.vigursii", "E.anglica")))

sim <- manyhost_manyparasite_dynamics(field.size = 100^2, 
   interaction.matrix = int.mat, 
   host.number = c(3000, 3000), 
   parasite.number = c(1, 1), 
   gens = 100, 
   reps = 10)

examples to come...
}
