% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Many_host_single_parasite_dynamics.R
\name{manyhost_oneparasite_dynamics}
\alias{manyhost_oneparasite_dynamics}
\title{Many hosts, single parasite simulation for many iterations.}
\usage{
manyhost_oneparasite_dynamics(field.size = 100^2, s = c(0, 0, 0, 0, 0,
  5, 6, 7, 8, 9), host.number = c(rep(10000/10, 10)),
  parasite.number = 1, gens = 100, reps = 100)
}
\arguments{
\item{field.size}{size of host parasite matrix to play on. Must be square.}

\item{s}{the selective advantage to a parasite of landing on a host. Supply a vector of values.}

\item{host.number}{number of hosts to seed the matrix with. Supply a vector of values.}

\item{parasite.number}{number of parasites to seed the matrix with.}

\item{gens}{number of generations to run the model for.}
}
\value{
a list of the parasite dynamics, spatial structure and the host matrix?
}
\description{
Take a host grid containing a number of hosts with some benefit to the parasite (randomly spaced, but fixed).
Then overlay a parasite grid contaning some parasites (randomly spaced).
If hosts and parasites co-occur, the parasite will reproduce. Cumulative offspring then randomly overlaid on grid.
}
\details{
Model assumes (1) that hosts are fixed in position and perennial, and are unaffected by parasitism.
  (2) no effect on parasite of nearby hosts, (3) parasites seed randomly over the field, (4) all parasites survive to end of generation,
  (5) parasites do not produce any offspring without a host, and all parasites with a host reproduce and (6) no host independent mortality of parasites
}
\examples{
# need ggplot
library(ggplot2)
sim <- manyhost_oneparasite_dynamics()
ggplot(sim, aes(x = generation, y = P.size, group = rep)) + geom_line()
}
