% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_host_single_parasite_dynamics.R
\name{onehost_oneparasite_dynamics}
\alias{onehost_oneparasite_dynamics}
\title{Single host, single parasite simulation for many iterations.}
\usage{
onehost_oneparasite_dynamics(field.size = 10^2, s = 3,
  host.number = 35, parasite.number = 20, gens = 100, reps = 100)
}
\arguments{
\item{field.size}{size of host parasite matrix to play on. Must be square.}

\item{s}{the selective advantage to a parasite of landing on a host.}

\item{host.number}{number of hosts to seed the matrix with.}

\item{parasite.number}{number of parasites to seed the matrix with.}

\item{gens}{number of generations to run the model for.}

\item{reps}{number of replicates of each simulation.}
}
\description{
Take a host grid containing a number of hosts with some benefit to the parasite (randomly spaced, but fixed).
Then overlay a parasite grid contaning some parasites (randomly spaced).
If hosts and parasites co-occur, the parasite will reproduce. Cumulative offspring then randomly overlaid on grid.
}
\details{
This function is a wrapper for onehost_oneparasite() and simply replicates the number of simulations nicely.

Model assumes (1) that hosts are fixed in position and perennial, and are unaffected by parasitism.
  (2) no effect on parasite of nearby hosts, (3) parasites seed randomly over the field, (4) all parasites survive to end of generation,
  (5) parasites do not produce any offspring without a host, and all parasites with a host reproduce and (6) no host independent mortality of parasites
}
\examples{

# should print out simulation from default parameters
onehost_oneparasite() 
}
\seealso{
onehost_oneparasite_dynamics to replicate many times.
}
